# backend/docs/openapi-3.yaml
components:
  schemas:
...
    api.LoginSuccessResponse:
      properties:
        message:
          type: string
        session:
          $ref: '#/components/schemas/api.SessionData'
        user:
          $ref: '#/components/schemas/api.UserPublicResponse'
      type: object
...
    api.PasswordChangeResponse:
      properties:
        message:
          type: string
        success:
          type: boolean
      type: object
...
    api.SessionData:
      properties:
        expiresAt:
          type: string
        refreshToken:
          type: string
        token:
          type: string
      type: object
    api.SessionRefreshResponse:
      properties:
        expiresAt:
          type: string
        message:
          type: string
      type: object
    api.StandardErrorResponse:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    api.UserPublicResponse:
      properties:
        email:
          type: string
        id:
          type: string
        isActive:
          type: boolean
        username:
          type: string
      type: object
...
    models.ChangePasswordRequest:
      properties:
        currentPassword:
          type: string
        newPassword:
          minLength: 12
          type: string
      required:
      - currentPassword
      - newPassword
      type: object
...
    models.LoginRequest:
      properties:
        captchaToken:
          type: string
        email:
          type: string
        password:
          minLength: 6
          type: string
        rememberMe:
          type: boolean
      required:
      - email
      - password
      type: object
...
    models.User:
      properties:
        avatarUrl:
          type: string
        createdAt:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
        firstName:
          type: string
        id:
          type: string
        isActive:
          type: boolean
        isLocked:
          type: boolean
        lastLoginAt:
          type: string
        lastLoginIp:
          items:
            type: integer
          type: array
          uniqueItems: false
        lastName:
          type: string
        mfaEnabled:
          description: MFA support fields
          type: boolean
        mfaLastUsedAt:
          type: string
        mustChangePassword:
          type: boolean
        name:
          description: Computed fields (not stored in DB)
          type: string
        updatedAt:
          type: string
      type: object
...
  securitySchemes:
    basic:
      scheme: basic
      type: http
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for domain generation, validation, and management
    operations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Domain Flow API
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/change-password:
    post:
      description: Change password for the currently authenticated user
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.ChangePasswordRequest'
        description: Password change request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PasswordChangeResponse'
          description: Password changed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid request format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Authentication required
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Not implemented
      summary: Change user password
      tags:
      - authentication
  /auth/login:
    post:
      description: Authenticate user credentials and create session
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.LoginRequest'
        description: Login credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.LoginSuccessResponse'
          description: Login successful with user and session info
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid request format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid credentials
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Account inactive
        "423":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Account locked
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Internal server error
      summary: User login
      tags:
      - authentication
...
  /auth/me:
    get:
      description: Get information about the currently authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.User'
          description: Current user information
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: User not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Internal server error
      summary: Get current user
      tags:
      - authentication
  /auth/refresh:
    post:
      description: Extend the current session expiry time
      operationId: refreshSession
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SessionRefreshResponse'
          description: Session refreshed with new expiry
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid or expired session
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Failed to refresh session
      summary: Refresh user session
      tags:
      - authentication
...
servers:
- url: /api/v2